Refactor: split CacheMap into a source file
